+++ source/resources/lib/puma/puma.config/statsd_instrument.rb
@@ -1,0 +1,16 @@
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
+#  with the License. A copy of the License is located at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES
+#  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions
+#  and limitations under the License.
+
+STATSD_ADDR='localhost:8125'
+STATSD_IMPLEMENTATION='statsd'
+STATSD_PREFIX='sample_app'
+STATSD_ENV='production'
+STATSD_FLUSH_INTERVAL='0.0'